delete Expression DSL operators and old Expression class
