+++ cpp/src/arrow/dataset/api.h
@@ -21,9 +21,9 @@
 
 #include "arrow/dataset/dataset.h"
 #include "arrow/dataset/discovery.h"
+#include "arrow/dataset/expression.h"
 #include "arrow/dataset/file_base.h"
 #include "arrow/dataset/file_csv.h"
 #include "arrow/dataset/file_ipc.h"
 #include "arrow/dataset/file_parquet.h"
-#include "arrow/dataset/filter.h"
 #include "arrow/dataset/scanner.h"
