+++ cpp/src/arrow/dataset/partition_test.cc
@@ -105,17 +105,20 @@ TEST_F(TestPartitioning, DirectoryPartitioning) {
   partitioning_ = std::make_shared<DirectoryPartitioning>(
       schema({field("alpha", int32()), field("beta", utf8())}));
 
-  AssertParse("/0/hello", "alpha"_ == int32_t(0) and "beta"_ == "hello");
-  AssertParse("/3", "alpha"_ == int32_t(3));
+  AssertParse("/0/hello", and_(equal(field_ref("alpha"), literal(0)),
+                               equal(field_ref("beta"), literal("hello"))));
+  AssertParse("/3", equal(field_ref("alpha"), literal(3)));
   AssertParseError("/world/0");    // reversed order
   AssertParseError("/0.0/foo");    // invalid alpha
   AssertParseError("/3.25");       // invalid alpha with missing beta
   AssertParse("", literal(true));  // no segments to parse
 
   // gotcha someday:
-  AssertParse("/0/dat.parquet", "alpha"_ == int32_t(0) and "beta"_ == "dat.parquet");
+  AssertParse("/0/dat.parquet", and_(equal(field_ref("alpha"), literal(0)),
+                                     equal(field_ref("beta"), literal("dat.parquet"))));
 
-  AssertParse("/0/foo/ignored=2341", "alpha"_ == int32_t(0) and "beta"_ == "foo");
+  AssertParse("/0/foo/ignored=2341", and_(equal(field_ref("alpha"), literal(0)),
+                                          equal(field_ref("beta"), literal("foo"))));
 }
 
 TEST_F(TestPartitioning, DirectoryPartitioningFormat) {
@@ -124,20 +127,28 @@ TEST_F(TestPartitioning, DirectoryPartitioningFormat) {
 
   written_schema_ = partitioning_->schema();
 
-  AssertFormat("alpha"_ == int32_t(0) and "beta"_ == "hello", "0/hello");
-  AssertFormat("beta"_ == "hello" and "alpha"_ == int32_t(0), "0/hello");
-  AssertFormat("alpha"_ == int32_t(0), "0");
-  AssertFormatError("beta"_ == "hello");
+  AssertFormat(and_(equal(field_ref("alpha"), literal(0)),
+                    equal(field_ref("beta"), literal("hello"))),
+               "0/hello");
+  AssertFormat(and_(equal(field_ref("beta"), literal("hello")),
+                    equal(field_ref("alpha"), literal(0))),
+               "0/hello");
+  AssertFormat(equal(field_ref("alpha"), literal(0)), "0");
+  AssertFormatError(equal(field_ref("beta"), literal("hello")));
   AssertFormat(literal(true), "");
 
   ASSERT_OK_AND_ASSIGN(written_schema_,
                        written_schema_->AddField(0, field("gamma", utf8())));
-  AssertFormat("gamma"_ == "yo" and "alpha"_ == int32_t(0) and "beta"_ == "hello",
+  AssertFormat(and_({equal(field_ref("gamma"), literal("yo")),
+                     equal(field_ref("alpha"), literal(0)),
+                     equal(field_ref("beta"), literal("hello"))}),
                "0/hello");
 
   // written_schema_ is incompatible with partitioning_'s schema
   written_schema_ = schema({field("alpha", utf8()), field("beta", utf8())});
-  AssertFormatError<StatusCode::TypeError>("alpha"_ == "0.0" and "beta"_ == "hello");
+  AssertFormatError<StatusCode::TypeError>(
+      and_(equal(field_ref("alpha"), literal("0.0")),
+           equal(field_ref("beta"), literal("hello"))));
 }
 
 TEST_F(TestPartitioning, DirectoryPartitioningWithTemporal) {
@@ -147,7 +158,9 @@ TEST_F(TestPartitioning, DirectoryPartitioningWithTemporal) {
 
     ASSERT_OK_AND_ASSIGN(auto day, StringScalar("2020-06-08").CastTo(temporal));
     AssertParse("/2020/06/2020-06-08",
-                "year"_ == int32_t(2020) and "month"_ == int8_t(6) and "day"_ == day);
+                and_({equal(field_ref("year"), literal(2020)),
+                      equal(field_ref("month"), literal<int8_t>(6)),
+                      equal(field_ref("day"), literal(day))}));
   }
 }
 
@@ -207,7 +220,7 @@ TEST_F(TestPartitioning, DictionaryHasUniqueValues) {
         std::make_shared<DictionaryScalar>(index_and_dictionary, alpha->type());
 
     auto path = "/" + expected_dictionary->GetString(i);
-    AssertParse(path, "alpha"_ == dictionary_scalar);
+    AssertParse(path, equal(field_ref("alpha"), literal(dictionary_scalar)));
   }
 
   AssertParseError("/yosemite");  // not in inspected dictionary
@@ -223,17 +236,21 @@ TEST_F(TestPartitioning, HivePartitioning) {
   partitioning_ = std::make_shared<HivePartitioning>(
       schema({field("alpha", int32()), field("beta", float32())}));
 
-  AssertParse("/alpha=0/beta=3.25", "alpha"_ == int32_t(0) and "beta"_ == 3.25f);
-  AssertParse("/beta=3.25/alpha=0", "beta"_ == 3.25f and "alpha"_ == int32_t(0));
-  AssertParse("/alpha=0", "alpha"_ == int32_t(0));
-  AssertParse("/beta=3.25", "beta"_ == 3.25f);
+  AssertParse("/alpha=0/beta=3.25", and_(equal(field_ref("alpha"), literal(0)),
+                                         equal(field_ref("beta"), literal(3.25f))));
+  AssertParse("/beta=3.25/alpha=0", and_(equal(field_ref("beta"), literal(3.25f)),
+                                         equal(field_ref("alpha"), literal(0))));
+  AssertParse("/alpha=0", equal(field_ref("alpha"), literal(0)));
+  AssertParse("/beta=3.25", equal(field_ref("beta"), literal(3.25f)));
   AssertParse("", literal(true));
 
   AssertParse("/alpha=0/unexpected/beta=3.25",
-              "alpha"_ == int32_t(0) and "beta"_ == 3.25f);
+              and_(equal(field_ref("alpha"), literal(0)),
+                   equal(field_ref("beta"), literal(3.25f))));
 
   AssertParse("/alpha=0/beta=3.25/ignored=2341",
-              "alpha"_ == int32_t(0) and "beta"_ == 3.25f);
+              and_(equal(field_ref("alpha"), literal(0)),
+                   equal(field_ref("beta"), literal(3.25f))));
 
   AssertParse("/ignored=2341", literal(true));
 
@@ -246,20 +263,28 @@ TEST_F(TestPartitioning, HivePartitioningFormat) {
 
   written_schema_ = partitioning_->schema();
 
-  AssertFormat("alpha"_ == int32_t(0) and "beta"_ == 3.25f, "alpha=0/beta=3.25");
-  AssertFormat("beta"_ == 3.25f and "alpha"_ == int32_t(0), "alpha=0/beta=3.25");
-  AssertFormat("alpha"_ == int32_t(0), "alpha=0");
-  AssertFormat("beta"_ == 3.25f, "alpha/beta=3.25");
+  AssertFormat(and_(equal(field_ref("alpha"), literal(0)),
+                    equal(field_ref("beta"), literal(3.25f))),
+               "alpha=0/beta=3.25");
+  AssertFormat(and_(equal(field_ref("beta"), literal(3.25f)),
+                    equal(field_ref("alpha"), literal(0))),
+               "alpha=0/beta=3.25");
+  AssertFormat(equal(field_ref("alpha"), literal(0)), "alpha=0");
+  AssertFormat(equal(field_ref("beta"), literal(3.25f)), "alpha/beta=3.25");
   AssertFormat(literal(true), "");
 
   ASSERT_OK_AND_ASSIGN(written_schema_,
                        written_schema_->AddField(0, field("gamma", utf8())));
-  AssertFormat("gamma"_ == "yo" and "alpha"_ == int32_t(0) and "beta"_ == 3.25f,
+  AssertFormat(and_({equal(field_ref("gamma"), literal("yo")),
+                     equal(field_ref("alpha"), literal(0)),
+                     equal(field_ref("beta"), literal(3.25f))}),
                "alpha=0/beta=3.25");
 
   // written_schema_ is incompatible with partitioning_'s schema
   written_schema_ = schema({field("alpha", utf8()), field("beta", utf8())});
-  AssertFormatError<StatusCode::TypeError>("alpha"_ == "0.0" and "beta"_ == "hello");
+  AssertFormatError<StatusCode::TypeError>(
+      and_(equal(field_ref("alpha"), literal("0.0")),
+           equal(field_ref("beta"), literal("hello"))));
 }
 
 TEST_F(TestPartitioning, DiscoverHiveSchema) {
@@ -326,7 +351,7 @@ TEST_F(TestPartitioning, HiveDictionaryHasUniqueValues) {
         std::make_shared<DictionaryScalar>(index_and_dictionary, alpha->type());
 
     auto path = "/alpha=" + expected_dictionary->GetString(i);
-    AssertParse(path, "alpha"_ == dictionary_scalar);
+    AssertParse(path, equal(field_ref("alpha"), literal(dictionary_scalar)));
   }
 
   AssertParseError("/alpha=yosemite");  // not in inspected dictionary
@@ -365,9 +390,12 @@ TEST_F(TestPartitioning, EtlThenHive) {
       });
 
   AssertParse("/1999/12/31/00/alpha=0/beta=3.25",
-              "year"_ == int16_t(1999) and "month"_ == int8_t(12) and
-                  "day"_ == int8_t(31) and "hour"_ == int8_t(0) and
-                  ("alpha"_ == int32_t(0) and "beta"_ == 3.25f));
+              and_({equal(field_ref("year"), literal<int16_t>(1999)),
+                    equal(field_ref("month"), literal<int8_t>(12)),
+                    equal(field_ref("day"), literal<int8_t>(31)),
+                    equal(field_ref("hour"), literal<int8_t>(0)),
+                    and_(equal(field_ref("alpha"), literal<int32_t>(0)),
+                         equal(field_ref("beta"), literal<float>(3.25f)))}));
 
   AssertParseError("/20X6/03/21/05/alpha=0/beta=3.25");
 }
@@ -382,7 +410,7 @@ TEST_F(TestPartitioning, Set) {
   auto schm = schema({field("x", int32())});
 
   // An adhoc partitioning which parses segments like "/x in [1 4 5]"
-  // into ("x"_ == 1 or "x"_ == 4 or "x"_ == 5)
+  // into (field_ref("x") == 1 or field_ref("x") == 4 or field_ref("x") == 5)
   partitioning_ = std::make_shared<FunctionPartitioning>(
       schm, [&](const std::string& path) -> Result<Expression2> {
         std::vector<Expression2> subexpressions;
@@ -402,18 +430,21 @@ TEST_F(TestPartitioning, Set) {
           }
 
           subexpressions.push_back(call("is_in", {field_ref(matches[1])},
-                                        compute::SetLookupOptions{ints(set), true}));
+                                        compute::SetLookupOptions{ints(set)}));
         }
         return and_(std::move(subexpressions));
       });
 
-  AssertParse("/x in [1]", "x"_.In(ints({1})));
-  AssertParse("/x in [1 4 5]", "x"_.In(ints({1, 4, 5})));
-  AssertParse("/x in []", "x"_.In(ints({})));
+  auto x_in = [&](std::vector<int32_t> set) {
+    return call("is_in", {field_ref("x")}, compute::SetLookupOptions{ints(set)});
+  };
+  AssertParse("/x in [1]", x_in({1}));
+  AssertParse("/x in [1 4 5]", x_in({1, 4, 5}));
+  AssertParse("/x in []", x_in({}));
 }
 
 // An adhoc partitioning which parses segments like "/x=[-3.25, 0.0)"
-// into ("x"_ >= -3.25 and "x" < 0.0)
+// into (field_ref("x") >= -3.25 and "x" < 0.0)
 class RangePartitioning : public Partitioning {
  public:
   explicit RangePartitioning(std::shared_ptr<Schema> s) : Partitioning(std::move(s)) {}
@@ -477,8 +508,12 @@ TEST_F(TestPartitioning, Range) {
       schema({field("x", float64()), field("y", float64()), field("z", float64())}));
 
   AssertParse("/x=[-1.5 0.0)/y=[0.0 1.5)/z=(1.5 3.0]",
-              ("x"_ >= -1.5 and "x"_ < 0.0) and ("y"_ >= 0.0 and "y"_ < 1.5) and
-                  ("z"_ > 1.5 and "z"_ <= 3.0));
+              and_({and_(greater_equal(field_ref("x"), literal(-1.5)),
+                         less(field_ref("x"), literal(0.0))),
+                    and_(greater_equal(field_ref("y"), literal(0.0)),
+                         less(field_ref("y"), literal(1.5))),
+                    and_(greater(field_ref("z"), literal(1.5)),
+                         less_equal(field_ref("z"), literal(3.0)))}));
 }
 
 TEST(TestStripPrefixAndFilename, Basic) {
@@ -495,6 +530,58 @@ TEST(TestStripPrefixAndFilename, Basic) {
   EXPECT_THAT(StripPrefixAndFilename(input, "/data"),
               testing::ElementsAre("year=2019", "year=2019/month=12",
                                    "year=2019/month=12/day=01"));
+}
+
+void AssertGrouping(const FieldVector& by_fields, const std::string& batch_json,
+                    const std::string& expected_json) {
+  FieldVector fields_with_ids = by_fields;
+  fields_with_ids.push_back(field("ids", list(int32())));
+  auto expected = ArrayFromJSON(struct_(fields_with_ids), expected_json);
+
+  FieldVector fields_with_id = by_fields;
+  fields_with_id.push_back(field("id", int32()));
+  auto batch = RecordBatchFromJSON(schema(fields_with_id), batch_json);
+
+  ASSERT_OK_AND_ASSIGN(auto by, batch->RemoveColumn(batch->num_columns() - 1)
+                                    .Map([](std::shared_ptr<RecordBatch> by) {
+                                      return by->ToStructArray();
+                                    }));
+
+  ASSERT_OK_AND_ASSIGN(auto groupings_and_values, MakeGroupings(*by));
+
+  auto groupings =
+      checked_pointer_cast<ListArray>(groupings_and_values->GetFieldByName("groupings"));
+
+  ASSERT_OK_AND_ASSIGN(std::shared_ptr<Array> grouped_ids,
+                       ApplyGroupings(*groupings, *batch->GetColumnByName("id")));
+
+  ArrayVector columns =
+      checked_cast<const StructArray&>(*groupings_and_values->GetFieldByName("values"))
+          .fields();
+  columns.push_back(grouped_ids);
+
+  ASSERT_OK_AND_ASSIGN(auto actual, StructArray::Make(columns, fields_with_ids));
+
+  AssertArraysEqual(*expected, *actual, /*verbose=*/true);
+}
+
+TEST(GroupTest, Basics) {
+  AssertGrouping({field("a", utf8()), field("b", int32())}, R"([
+    {"a": "ex",  "b": 0, "id": 0},
+    {"a": "ex",  "b": 0, "id": 1},
+    {"a": "why", "b": 0, "id": 2},
+    {"a": "ex",  "b": 1, "id": 3},
+    {"a": "why", "b": 0, "id": 4},
+    {"a": "ex",  "b": 1, "id": 5},
+    {"a": "ex",  "b": 0, "id": 6},
+    {"a": "why", "b": 1, "id": 7}
+  ])",
+                 R"([
+    {"a": "ex",  "b": 0, "ids": [0, 1, 6]},
+    {"a": "why", "b": 0, "ids": [2, 4]},
+    {"a": "ex",  "b": 1, "ids": [3, 5]},
+    {"a": "why", "b": 1, "ids": [7]}
+  ])");
 }
 
 }  // namespace dataset
